datasets-cardiovasculares/cloud_functions/Riesgo/requirements.txt
Flask==2.0.3
pandas==1.3.3
numpy==1.26.4
joblib==1.4.2
scikit-learn==1.6.1
google-cloud-storage==2.2.1
Werkzeug==2.0.3
gunicorn==22.0.0
xgboost==2.1.4